Bug 1211699 - Escape invalid URL characters sent by the context menu helper
